+++ modules/ecs_service/service.tf
@@ -12,7 +12,7 @@ resource "aws_ecs_service" "main" {
   name             = var.name
   cluster          = data.aws_ecs_cluster.main.id
   task_definition  = aws_ecs_task_definition.main.arn
-  desired_count    = 1
+  desired_count    = var.replicas
   launch_type      = var.type
   platform_version = var.type == "FARGATE" ? "1.4.0" : null